Fix warning in optimize.brentq
Ultimatly this was because fzhang_opt was returning NaN Re was
evaluated as a negative number. This came from the velocity
(in u/mu) having a sign. Re should just depend on the speed of
the sphere, not the direction in which it is moving so take the
magnitude of u when calculating Re everywhere.

The results all look the same. The diff shows some (very small)
differences in output. I don't think this is a problem. It looks
like this is caused by brentq taking a sligtly different path to
an optimum solution (as it does not avoid those regions where it
sees NaNs).