Compile the inner loop for CSD update
This speeds things up from ~0.02 s / timestep
to ~0.00016 s / timestep. Worth having. Note that
I've had to remove an assert statment that checks
for us failing to find a new homw for a growing
crystal. Should probably check the number of input
particles matches the number of output particles
somewhere.