BL Fixes and Particle Calc
Fixed a bunch of minor issues with the BL analysis.  The main thing here is that I defined a new regime at very low Re and Pe where the boundary layers must scale as the particle radius. I then implemented the equations for estimating particle size and radius and a wrapper that iterates over viscosity and ICB growth rate.