Add note about restricted nuc, and find a bug
Add note about two ways we could think about the restricted
volume problem (and why I have chosen one of them) and output
composition of residual liquid. It turns out that this shows
an error somewhere in my calculation of the composition of
the residual liquid. It gets more Fe rich as we make the box
bigger. Critical radius does not change much, so something
seems wrong. Need to check calculation of molar fractions :(