Avoid spurious warnings
It seems that jitted functions being
called from python functions that are then
numpy vectorised leave some kind of floating
point exception flag set which numpy flags as
a warning. This slows things down and is annoying.
Short term fix is to patch the vectorised functions
to ignore these warnings. See
https://github.com/andreww/slurry/issues/8

Probably need to raise an issue in numpy or numba
or both :(