Work on proper self consistent solution
Today's little bit of fun was working out that some particles will
want to disolve (e.g in this setup those that form within 1 km of
the top of the F-layer. End up handling this with a new termination
event. These particles won't contribute to our integral over zn,
but we should be able to handle them gracefully (and thus deal with
Mars).