Working self consistent solution
This makes it possible to pass the region of interest radius
as a function and uses this in a self consistent solution for
the F-layer snowing problem. However, this way of doing it is
very slow (we are not using any information about the seperation
in the update to the minimiser for the self consistent soluton).
We may be able to do better as the solution is so insensitive to the
seperation in the end.