Numerical approximation to inverse Vinet EOS
This uses SimPy to compute the inverse of the third order
Taylor expansion of the Vinet EOS (see Etter and Dinnebier;
J. App. Crys. 2013). Works but it takes ages to solve and
I cannot get the lambdered solution to jit compile. Let's
just leave this in the repo as a historical artifact.