Remove the 'interaction sphere' from the IVP for a falling particle
We now just impose the composition at the outside of the boundary layer
and solve for the composition and gradient of the boundary layer to
match that. This means we don't impose the total composition of
a sphere around the particle but we can fix this by wrapping another
self consistent loop around everything to conserve the overall composition.