Remove old notebooks
We don't need this hand-written finite difference solver
to find the falling time of a particle, or the old notes
and code for the falling velocity (which is all in terms of
equilibrium timescale) so remove.

There are some notes in falling_timescale.ipynb that we may
want to look at to check thermal equilibration time, but these
can easily be recovered.