Another speed up. Now to 16s per F-layer case
This moves the gravity calculation for the pressure
into numba saving lots of python function calls. A
factor of three speedup and we now have about the
same time spend in PREM as in the thermodynamics.

There are ways to speed up both of these things, but
this is going to be fast enough I suspect.