Put nucleation inside flayer solver
Each iteration we recompute the nuclation rates and critical
radii at each integration depth. This goes to zero as we approach
equilibrium (increase O in the liquid or increase T towards the
liquidus). To stop us having massive initial radii (like 20 m)
we set a threahold minimum nucleation rate. If the layer has a
rate below this we do not solve the IVP and the layer contributes
no solid volume to the overall solution.