Force x, p and t to be scalars for calculation of nuclation, so we can numba
To work out what was going on force some of the numba jitting to be in
nopython mode (turns ugly warnings into a runtime crash). Shows that the
vectorised x, p and t functions of the f-layer return 0D arrays for when
evaluated in one place (grrr) and this breaks the numba. Wrap in floats
to enable number to do its stuff.

This takes us to an interesting new place. The jitted functions end up
giving us nan instead of a very low nucleation rate and very large
critical radius (overflow I think) which puts the IVP solver at the
top of the f-layer in a position of running (and crashing - r0 of nan
isn't going to work). Fix this by treating nuc rate of nan as nuc rate
of zero, not running the solution, and not including non-nucleated
particles in the integral.