Work on automating run for many cases
The idea is to wrap each case in a dict, append the
results to the dict and save it. Use **dict to pass in
arguments. This needed some new output data from the
case solver in flayer, harmonising argument names,
and a minor change to the single case notebook to
match.