Avoid evaluating gravity and pressure for R<0
The ODE solver sometimes needs to find a solution
for positions far beyond the ICB in order to bracket
the termination of the run. If a particle is falling
very quickly this can result in an attempt to solve
for the falling velocity and growth rate at negative
radius (from the center of the Earth). Such results
never make it into out solutions but have to give
a reasonable answer so that the time when the ICB
is hit can be found.

This led to several problems, specifically triggering
of a bare assert in the code that finds the coefficents
for the gravity in the PREM module. This change:

* Turns this assert into something more helpful in a
  way that works with numba
* Handles the r=0 case properly
* Modifies the pressure and gravity functions passed
  into the solver such that they give P(r=0) and g(r=0)
  for any case where r < 0.
* Avoid the resulting division by zero in the scalar
  evaluation of g at the center of the Earth (just return
  g(r=0) = 0) in the same way as this was handled for the
  'vectorised' version.

Some of these changes should probably be moved out into the
PREM package once that gets wrapped up.