Changes to nucleation code for flayer model
In some cases we would hit the liquidus exactly. This falied
and gave division by zero (as gsl is zero). Avoid this by
checking sign of gsl and if it is zero of positive (on or
above liquidus) just return nan for all parameters.

It's also useful to know the critical energy in the flayer code,
so expose this as a result from the nuc stuff. Does not change
what we plot but involes adding output argumets and catching these
in the nucleation notebook and the flayer model.

Finally, tweak the plotting in the nucleation notebook to avoind
a tight layout warning. Just a slight formatting difference. Add
(a) and (b) to the plots.