Add informational output when the falling velocity solver fails
A common error is that the bounds of the self consistent solution for the
velocity (and thus Re) have the same sign. This leads to a crash but not
much useful info. This adds explicit calls before the optimisation to check
the sign and, if they are the same, print all input before calling the
optimiser (which crashes).