Fix to run on newest python on Andrew's laptop
We still see the warnings about invalid values for numba and
numpy vectorized functions. This is, I think, ultimatly a bug in
the underlying vector operations that means that input that never
gets fed into an operation that gives an invalid input still ends
up flagging the invalid bit (which gets picked up and turned into
a python warning). This is all "fine" except we filter warnings out
of the Re solver to exclude solutions that are outside the C_D domain.
Fix this by adding the UserWarning class to the warning filter that
turns these into errors.