A note pointing out where we may want to look to improve convergence
I suspect we can add a test in here to use (and account for) a
larger tau if we step outside the domain of the IVP solution. But
first check what actually happens when we step outside of the IVP
solution. Is this the cause of the poor scaling for CSD?