Strip out calculation of thermal boundary layer thickness
The only functional change is in falling.py (in the boundary layer
calculation code, which I'm about to change, which is why removing
this is useful) but it ends up changing the output from the overall
particle dynamics wrapper (self consistent velocity and BL thickness)
hence changing the particle evolution code where this is used. Nice that
it does not propogate further into the F-layer code, for example.

Also add an exrra line for Fr = 5 and Fr=15 in the falling plot. This
will show where we interpolate, nicly between the two chosen cases
in the figure.