Handle changing Fr regime at int or high Re
We did end up with a discontiuity in the boundary layer thickness
if we moved from high to low Fr (which could happen in some cases
for particles around the edge of 'sensible' parameters). Now
do linear interpolartion for between high and low Fr if Fr is
between 5 and 50 (the transition should be at Fr=10).