Introduce hetrogeniuos nuc to the nucleation module
Add optional argumet theta, the wetting angle in degrees.
Default of 180.0 gives homogenious nucleation. About 10
degrees should correspond to metal seeds, lowers the
critical energy and increases the nucleation rate. However,
this approach (see Christian) defines rc as the critical
radius for a sphere where the spherical cap sticking out
above a flat surface is the thing that forms. It's not
clear this this is the correct radius to imagine as the
starting size of our snow. Should we have a 'seed radius'
parameter too to initialise the IVPs?