New panel for growth figure showing interface composition
This is the other peice of the solution that is useful for explaining what is going on.
Basically, we end up with slow growth that gives liquid composition at equilibrium. And
start with fast growth where the compositon is the same as the bulk liquid, and thus
gives the maximum growth rate.