Avoid optimizing the upper boundary layer temperature
This stops the upper boundary condition temperature being
a parameter that the optimizer can change. Without this
we can get the (valid) high T (above Tm) isothermal solution
with no solid and no heat flux. However, this solution breakes
the 'top of f-layer is on Tm) assumption. So avoid this.

With this change we seem to get a converged solution.