Build and explore a new temperature function
This ends up being a closure that returns a factory. Hum, somebody
should have a word with the archetect. Anyway, we end up with a way
to pass in n_points - 1 parameters, the first is a gradient, and the
others are temperature putubations at each interior point on our grid.
The bottom does not change (no point, just change the gradient,
something needs to do this), nor does the top (tied to liquidus).

In principle I could lose another parameter and impose the zero
flux lower boundary condition (but we may want to turn that off
for direct crystalisation) or have a smooth curve through the points
(because the intermediate values matter in the solution) but let's
test this first.