Harden the single particle calculation
The IVP solver needs to be able to evaluate the growth
rate and falling velocity in all sorts of odd places, so
we can easily get things like very high temperatures inside
the inner core or negative particle radii. These end up
not mattering in the calculation. Here we avoid this kind
of thing crashing the code (which seems to happen more if
we have finer grids of points).