Just some comments to help me understand c diff
So the issue is that the compositional diffusion
solver fails with rather low mass production rates
if we include the oxygen composition function in
the source. Taking it out (and solving this in
the outer loop) fixes things until we get high
enough production to go unreasonably high in
oxygen content in the liquid.