N_BV calculator in the notebook
This also shows how hard it is to get a stable
liquid layer if we relase oxygen and heat. The
effect of oxygen on the liquid density is so
large that you need a big subadiabatic cooling
to end up with denser material at the botom of
the layer than the adiabatic reference.