Put Brunt Vasala frequency in particle evolution
We have been using a very small (but real) BV frequency
in all calculations (other than the particle evolution
notebook). Instead pass this in and use it in calculation
of Fr and thus boundary layer thickness. The BV frequency
must come from outside the particle evolution solver (as
this does not know everything needed for it's calculation)
so pass it in (and do the calculation outside).

Add a warning if this is missing (but most code paths also
make this compusory).