Calculate and output information on excess density
This is the density difference between the model (just
above the ICB) and the projected adiabatic profile. There
are two components, the liquid (which is reported as a difference in density
between the adiabat and the liquid and is calculated alongside
the BV frequencey), and the solid (which is reported as the
difference in density between the pure liquid and the liquid
and solid slurry). Adding the two together gives the total
difference between the adiabat and the layer. If we take
prem as a measure of the adiabat and AK135 as the measure
for the Earth, the total excess density should be 27.24
kg/m^3. In practice it seems that this is dominated by the
liquid (low solid volume fraction).