Fix two errors discovered in review
One of our referees noted a disagreement between
the calculated latent heat released by particle
growth in the F-layer and the growth rate of the
inner core by sedimentation which can be converted
into a latent heat release rate. The growth rate
was about half as big as it should have been (or
the latent heat release was a factor of two too
big). It turns out that this can be resolved by
fixing two errors both relating to the fact that
half of the particles that reach the critical
radius remelt:

* First, the factor of 1/2 (or I(R)/2) in equation
5.9 was missing from the code. This has been added.

* Second, the factor of 1/2 in the waiting time
calculation (equation 4.5, which comes from eq 6
of Davies et al. 2019) is upside down because if
we reduce the rate of nucleation we should increase
the time between nucleation events. This is also
corrected.

Overall this resolves the disagreement by halving
the growth rate and quartering the latent heat
release.